🚧 feat: 피드 삭제 버튼 눌렀을 시 확인하는 window.confirm 추가

+++ src/components/Feed/Question.jsx
@@ -30,9 +30,10 @@ const List = ({ questions, subjectData }) => {
   const { pathname } = useLocation();
 
   const handleAllDelete = async () => {
-    console.log(subjectId);
-    await deleteSubject({ subjectId });
-    navigate('/list');
+    if (window.confirm('정말 피드를 삭제하시겠습니까?')) {
+      await deleteSubject({ subjectId });
+      navigate('/list');
+    }
   };
 
   useEffect(() => {
